+++ SRC/Private/Extensions/TypeExtensions.cs
@@ -5,11 +5,8 @@
 ********************************************************************************/
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
-using System.Text.RegularExpressions;
 
 namespace Solti.Utils.DI.Internals
 {
@@ -85,92 +82,6 @@ namespace Solti.Utils.DI.Internals
                 throw new ArgumentException(Resources.CONSTRUCTOR_NOT_FOUND, nameof(argTypes));
 
             return ctor.Call(args);
-        }
-
-        public static IReadOnlyList<Type> GetOwnGenericArguments(this Type src) => src
-            .GetGenericArguments()
-            .Except(src.DeclaringType?.GetGenericArguments() ?? Array.Empty<Type>(), new GenericArgumentComparer())
-            .ToArray();
-
-        //
-        // Sajat comparer azert kell mert pl List<T> es IList<T> eseten typeof(List<T>).GetGenericArguments[0] != typeof(IList<T>).GetGenericArguments[0] 
-        // testzoleges "T"-re.
-        //
-
-        private sealed class GenericArgumentComparer : IEqualityComparer<Type>
-        {
-            public bool Equals(Type x, Type y) => GetHashCode(x) == GetHashCode(y);
-
-            //
-            // Generikus argumentumnak nincs teljes neve ezert a lenti sor jol kezeli a fenti
-            // problemat.
-            //
-
-            [SuppressMessage("Globalization", "CA1307: Specify StringComparison", Justification = "Type names are not localized.")]
-            public int GetHashCode(Type obj) => (obj.FullName ?? obj.Name).GetHashCode();
-        }
-
-        public static IReadOnlyList<Type> GetParents(this Type type)
-        {
-            return GetParentsInternal().Reverse().ToArray();
-
-            IEnumerable<Type> GetParentsInternal()
-            {
-                for (Type parent = type.DeclaringType; parent != null; parent = parent.DeclaringType)
-                    yield return parent;
-            }
-        }
-
-        //
-        // "&":  referencia szerinti parameter
-        // "`d": generikus tipus ahol "d" egesz szam
-        // "[T, TT]": generikus parameterek
-        //
-
-        private static readonly Regex TypeNameReplacer = new Regex(@"\&|`\d+(\[[\w,]+\])?", RegexOptions.Compiled);
-
-        public static string GetFriendlyName(this Type src)
-        {
-            Debug.Assert(!src.IsGenericType() || src.IsGenericTypeDefinition());
-            return TypeNameReplacer.Replace(src.IsNested ? src.Name : src.ToString(), string.Empty);
-        }
-
-        public static IEnumerable<TMember> ListMembers<TMember>(this Type src, Func<Type, BindingFlags, TMember[]> factory, bool includeNonPublic = false) where TMember : MemberInfo
-        {
-            BindingFlags flags = BindingFlags.Public | BindingFlags.Instance;
-
-            if (src.IsInterface())
-                return factory(src, flags)
-                    .Concat
-                    (
-                        src.GetInterfaces().SelectMany(iface => factory(iface, flags))
-                    );
-
-            //
-            // A "BindingFlags.NonPublic" es "BindingFlags.FlattenHierarchy" nem ertelmezett interface-ekre.
-            //
-
-            flags |= BindingFlags.FlattenHierarchy;
-            if (includeNonPublic) flags |= BindingFlags.NonPublic;
-
-            return factory(src, flags);
-        }
-
-        public static IEnumerable<Assembly> GetReferences(this Type src)
-        {
-            Assembly declaringAsm = src.Assembly();
-
-            var references = new[] { declaringAsm }.Concat(declaringAsm.GetReferences());
-
-            //
-            // Generikus parameterek szerepelhetnek masik szerelvenyben.
-            //
-
-            if (src.IsGenericType())
-                foreach (Type type in src.GetGenericArguments().Where(t => !t.IsGenericParameter))
-                    references = references.Concat(type.GetReferences());
-
-            return references.Distinct();
         }
     }
 }